requirements.txt
Flask
line-bot-sdk
pymongo
openai==1.30.0